requirements.txt
python-dotenv>=0.19.0
python-telegram-bot
openai>=1.0